Creating new presenter so that app and autofill can both inherit from RoutePresenter.
